Added cllicking function
Used chat GPT to add in a clicking function, which I can rework to make my own, and helps me to understand what to do. Also added in a testing page, so that I can test and figure out aspects before adding them in. E.g the nav bar which im trying to fix